requirements.txt
# Core
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
pyyaml>=6.0

# Machine Learning
scikit-learn>=1.0.0
imbalanced-learn>=0.9.0
xgboost>=1.6.0

# Deep Learning
torch>=1.12.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0

# Progress
tqdm>=4.64.0

# Optional: GPU acceleration
# torch with CUDA support - install from https://pytorch.org/
